requirements.txt
azure-functions
azure-identity
feedparser>=6.0.0
requests>=2.25.0
python-dotenv>=0.19.0
msal
babel>=2.17.0